+++ tests/qlatest.cpp
@@ -145,7 +145,7 @@ bool TestEncoders(int curLine, AmpIO &Board, BasePort *Port, std::ofstream &logF
     unsigned long darray[4];
     unsigned long testValue[4];
     char numStr[4][3] = { " 1", " 2", " 3", " 4" };
-    char buf[80];
+    char buf[250];
     bool pass = true;
     bool tmpFix;
 
@@ -653,7 +653,11 @@ bool TestPowerAmplifier(int curLine, AmpIO &Board, BasePort *Port, std::ofstream
         for (i = 0; i < 4; i++) {
             cur[i] = Board.GetMotorCurrent(i);
             logFile << std::hex << cur[i];
-            if (abs(cur[i]-dac) > 0x0100) {
+            if ((cur[i] > (dac + 0x0100))
+                || ((dac > 0x0100)
+                    && (cur[i] < (dac - 0x0100))
+                    )
+                ) {
                 logFile << "* ";
                 mvprintw(curLine+2, 15+7*i, "FAIL");
                 pass = false;