+++ CHANGELOG.md
@@ -1,6 +1,26 @@
 Change log
 ==========
 
+1.5.0 (2019-04-19)
+==================
+* API changes:
+  * None
+* New features:
+  * Support FPGA/QLA serial numbers with up to 3 digits
+  * Support to read Dallas chip on da Vinci tools (requires firmware 7 to be released)
+* Bug fixes:
+  * None
+
+1.4.0 (2018-05-16)
+==================
+* API changes:
+  * None
+* New features:
+  * Support for firmware Rev6 velocity estimation
+  * `pgm1394` now displays version for AmpIO library
+* Bug fixes:
+  * Fixed CMakeLists project names
+
 1.3.0 (2017-11-07)
 ==================
 * API changes: